products/requirements.txt
aniso8601==9.0.1
asgiref==3.5.0
backports.zoneinfo==0.2.1
black==22.1.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.0.4
Django==4.0.4
django-ckeditor==6.4.1
django-ckeditor-5==0.1.0
django-debug-toolbar==3.4.0
django-js-asset==2.0.0
djangorestframework==3.13.1
Flask==2.1.2
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.1
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
